pyproject.toml
[tool.poetry]
name = "socraticai"
version = "0.1.0"
description = "A CLI tool for AI-powered audio transcription and analysis"
authors = ["Ian Eisenberg <IanEisenberg90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
en_core_web_lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0.tar.gz" }
python = "^3.11"
numpy = "^2.0.0"
pandas = "^2.2.3"
scipy = "^1.14.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scikit-learn = "^1.5.2"
nltk = "^3.9.1"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"
jupyterlab = "^4.3.1"
fastapi = "^0.115.5"
anthropic = "0.49.0"
assemblyai = "^0.35.1"
tiktoken = "^0.8.0"
python-dotenv = "^1.0.1"
openai = "^1.54.5"
spacy = "^3.8.2"
tenacity = "^9.0.0"
pyyaml = "^6.0.1"
pyperclip = "^1.9.0"
google-genai = "^1.16.1"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
socraticai = "cli.main:cli"

[tool.pytest.ini_options]
# Test discovery
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

# Custom markers
markers = [
    "unit: marks tests as unit tests (fast, no external dependencies)",
    "integration: marks tests as integration tests (may make real API calls)",
    "slow: marks tests as slow running",
]

# Default behavior
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]

# Logging
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"